// test/integration/test/ports/engines/queries/filters/many_relation/_fixture/23742/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Bottom: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly middleId: CodecTypes['pg/int4@1']['output'] | null;
    };
    readonly Middle: { readonly id: CodecTypes['pg/int4@1']['output'] };
    readonly Top: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly middleId: CodecTypes['pg/int4@1']['output'] | null;
    };
    readonly TopBottom: {
      readonly topId: CodecTypes['pg/int4@1']['output'];
      readonly bottomId: CodecTypes['pg/int4@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Bottom: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly middleId: CodecTypes['pg/int4@1']['input'] | null;
    };
    readonly Middle: { readonly id: CodecTypes['pg/int4@1']['input'] };
    readonly Top: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly middleId: CodecTypes['pg/int4@1']['input'] | null;
    };
    readonly TopBottom: {
      readonly topId: CodecTypes['pg/int4@1']['input'];
      readonly bottomId: CodecTypes['pg/int4@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly bottom: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly middleId: CodecTypes['pg/int4@1']['output'] | null;
    };
    readonly middle: { readonly id: CodecTypes['pg/int4@1']['output'] };
    readonly top: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly middleId: CodecTypes['pg/int4@1']['output'] | null;
    };
    readonly topBottom: {
      readonly bottomId: CodecTypes['pg/int4@1']['output'];
      readonly topId: CodecTypes['pg/int4@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly bottom: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly middleId: CodecTypes['pg/int4@1']['input'] | null;
    };
    readonly middle: { readonly id: CodecTypes['pg/int4@1']['input'] };
    readonly top: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly middleId: CodecTypes['pg/int4@1']['input'] | null;
    };
    readonly topBottom: {
      readonly bottomId: CodecTypes['pg/int4@1']['input'];
      readonly topId: CodecTypes['pg/int4@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;