// test/integration/test/ports/engines/queries/filters/many_relation/_fixture/25104/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly A: { readonly id: CodecTypes['pg/text@1']['output'] };
    readonly B: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly aId: CodecTypes['pg/text@1']['output'];
    };
    readonly BC: {
      readonly bId: CodecTypes['pg/text@1']['output'];
      readonly cId: CodecTypes['pg/text@1']['output'];
    };
    readonly C: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly A: { readonly id: CodecTypes['pg/text@1']['input'] };
    readonly B: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly aId: CodecTypes['pg/text@1']['input'];
    };
    readonly BC: {
      readonly bId: CodecTypes['pg/text@1']['input'];
      readonly cId: CodecTypes['pg/text@1']['input'];
    };
    readonly C: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly a: { readonly id: CodecTypes['pg/text@1']['output'] };
    readonly b: {
      readonly aId: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
    };
    readonly bC: {
      readonly bId: CodecTypes['pg/text@1']['output'];
      readonly cId: CodecTypes['pg/text@1']['output'];
    };
    readonly c: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly a: { readonly id: CodecTypes['pg/text@1']['input'] };
    readonly b: {
      readonly aId: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
    };
    readonly bC: {
      readonly bId: CodecTypes['pg/text@1']['input'];
      readonly cId: CodecTypes['pg/text@1']['input'];
    };
    readonly c: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;