// test/integration/test/ports/engines/queries/filters/many_relation/_fixture/common/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Blog: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Comment: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly text: CodecTypes['pg/text@1']['output'];
      readonly likes: CodecTypes['pg/int4@1']['output'];
      readonly post_id: CodecTypes['pg/text@1']['output'];
    };
    readonly Post: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly popularity: CodecTypes['pg/int4@1']['output'];
      readonly blog_id: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Blog: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Comment: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly text: CodecTypes['pg/text@1']['input'];
      readonly likes: CodecTypes['pg/int4@1']['input'];
      readonly post_id: CodecTypes['pg/text@1']['input'];
    };
    readonly Post: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly popularity: CodecTypes['pg/int4@1']['input'];
      readonly blog_id: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly blog: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly comment: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly likes: CodecTypes['pg/int4@1']['output'];
      readonly post_id: CodecTypes['pg/text@1']['output'];
      readonly text: CodecTypes['pg/text@1']['output'];
    };
    readonly post: {
      readonly blog_id: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly popularity: CodecTypes['pg/int4@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly blog: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly comment: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly likes: CodecTypes['pg/int4@1']['input'];
      readonly post_id: CodecTypes['pg/text@1']['input'];
      readonly text: CodecTypes['pg/text@1']['input'];
    };
    readonly post: {
      readonly blog_id: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly popularity: CodecTypes['pg/int4@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;