// test/integration/test/ports/engines/queries/filters/many_relation/_fixture/different-pk/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Bottom: {
      readonly bottomId: CodecTypes['pg/int4@1']['output'];
      readonly relatedMiddleId: CodecTypes['pg/int4@1']['output'] | null;
    };
    readonly Middle: { readonly middleId: CodecTypes['pg/int4@1']['output'] };
    readonly Top: {
      readonly topId: CodecTypes['pg/int4@1']['output'];
      readonly relatedMiddleId: CodecTypes['pg/int4@1']['output'] | null;
    };
    readonly TopBottom: {
      readonly topId: CodecTypes['pg/int4@1']['output'];
      readonly bottomId: CodecTypes['pg/int4@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Bottom: {
      readonly bottomId: CodecTypes['pg/int4@1']['input'];
      readonly relatedMiddleId: CodecTypes['pg/int4@1']['input'] | null;
    };
    readonly Middle: { readonly middleId: CodecTypes['pg/int4@1']['input'] };
    readonly Top: {
      readonly topId: CodecTypes['pg/int4@1']['input'];
      readonly relatedMiddleId: CodecTypes['pg/int4@1']['input'] | null;
    };
    readonly TopBottom: {
      readonly topId: CodecTypes['pg/int4@1']['input'];
      readonly bottomId: CodecTypes['pg/int4@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly bottom: {
      readonly bottomId: CodecTypes['pg/int4@1']['output'];
      readonly relatedMiddleId: CodecTypes['pg/int4@1']['output'] | null;
    };
    readonly middle: { readonly middleId: CodecTypes['pg/int4@1']['output'] };
    readonly top: {
      readonly relatedMiddleId: CodecTypes['pg/int4@1']['output'] | null;
      readonly topId: CodecTypes['pg/int4@1']['output'];
    };
    readonly topBottom: {
      readonly bottomId: CodecTypes['pg/int4@1']['output'];
      readonly topId: CodecTypes['pg/int4@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly bottom: {
      readonly bottomId: CodecTypes['pg/int4@1']['input'];
      readonly relatedMiddleId: CodecTypes['pg/int4@1']['input'] | null;
    };
    readonly middle: { readonly middleId: CodecTypes['pg/int4@1']['input'] };
    readonly top: {
      readonly relatedMiddleId: CodecTypes['pg/int4@1']['input'] | null;
      readonly topId: CodecTypes['pg/int4@1']['input'];
    };
    readonly topBottom: {
      readonly bottomId: CodecTypes['pg/int4@1']['input'];
      readonly topId: CodecTypes['pg/int4@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;